odometry: send int's

+++ src/blackmetal/src/communication/controlSW/BlackMetal.cpp
@@ -38,8 +38,10 @@ void BlackMetal::onTwistRecievedSendJson(const geometry_msgs::msg::Twist &msg)
 	DBG("Message geometry_msgs::msg::Twist: " << geometry_msgs::msg::to_yaml(msg));
 	m_rightWheelSpeed = (msg.linear.x + 0.5 * m_chassisLength * msg.angular.z)/m_wheelRadius;
 	m_leftWheelSpeed = (msg.linear.x - 0.5 * m_chassisLength * msg.angular.z)/m_wheelRadius;
-	INFO("Right wheel speed: " << m_rightWheelSpeed << " Left wheel speed: " << m_leftWheelSpeed);
-	request(m_rightWheelSpeed, m_leftWheelSpeed);
+	INFO("Right wheel speed: " << static_cast<int>(m_rightWheelSpeed)
+	  << " Left wheel speed: " << static_cast<int>(m_leftWheelSpeed));
+	request(static_cast<int>(m_rightWheelSpeed),
+			static_cast<int>(m_leftWheelSpeed));
 }
 
 bm::Status BlackMetal::evalReturnState(const std::string &returnJson)
